This PR introduces a number of changes to the radonfourier cuda implementations based on discussions in https://github.com/PyLops/pylops/pull/611

Those changes improve the speed of both the 2D and 3D implementations (see https://github.com/PyLops/pylops_notebooks/blob/master/developement-cupy/Radon-Fourier.ipynb):

2D: forward 10.2ms, adjoint 11.2ms  --> forward 2.5ms, adjoint 2.4ms

3D: forward 9.8ms, adjoint 6.1ms  --> forward 3.1ms, adjoint 1.65ms

I would like to get this changes in before moving on to adding the grid-strided loops (https://github.com/PyLops/pylops/issues/616)