@mrava87 2024-12-11
@cako I am merging this so I can do a new release of PyLops. If you want to make any change feel free to open a new PR.